+++ site/src/app/routes/_page/demo/index.tsx
@@ -1,0 +1,6 @@
+import { createFileRoute } from "@tanstack/react-router";
+import { DocumentDemoRoute } from "../../../../routes/document-demo/DocumentDemoRoute";
+
+export const Route = createFileRoute("/_page/demo/")({
+  component: DocumentDemoRoute,
+});